requirements.txt
ultralytics
opencv-python
pyyaml
numpy
yt-dlp
pandas
matplotlib
seaborn
python-dateutil
fastapi
uvicorn[standard]
python-jose[cryptography]
overrides
cvzone
sqlalchemy[asyncio]
langchain
langchain_google_genai
langgraph
asyncpg
chromadb 
sentence-transformers 
google-generativeai 
python-docx
pydantic 
langchain-core
google-auth 
google-auth-oauthlib 
google-auth-httplib2
torch